Migrate to .Net Core #310 - some refactoring.
